Split dev API endpoint into separate router. Organize main API router
- Group all the API endpoint functions at the top of the api router file
- Move /dev/chat into a separate router and turn it on/off based on
  debug during app configuration
